+++ http/roadmap/RoadmapControllerHttpRequest.http
@@ -1,0 +1,6 @@
+### 내 과목 조회하기 (전체 조회 포함)
+GET {{host_url}}/v1/lectures/overviews?
+    page={{roadmap.API_3_6.page}}&
+    size={{roadmap.API_3_6.size}}
+Authorization: Bearer {{access_token}}
+Content-Type: application/json
