--- src/main/java/org/dongguk/onroad/loadmap/domain/CheckPoint.java
@@ -1,53 +1,0 @@
-package org.dongguk.onroad.loadmap.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "check_points")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class CheckPoint {
-
-    /* -------------------------------------------- */
-    /* Default Column ----------------------------- */
-    /* -------------------------------------------- */
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    /* -------------------------------------------- */
-    /* Information Column ------------------------- */
-    /* -------------------------------------------- */
-    @Column(name = "content", length = 500, nullable = false)
-    private String content;
-
-    /* -------------------------------------------- */
-    /* Timestamp Column --------------------------- */
-    /* -------------------------------------------- */
-    @Column(name = "created_at", nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    /* -------------------------------------------- */
-    /* Many To One Mapping ------------------------ */
-    /* -------------------------------------------- */
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "subtopic_id", nullable = false)
-    private Subtopic subtopic;
-
-    /* -------------------------------------------- */
-    /* Methods ------------------------------------ */
-    /* -------------------------------------------- */
-    @Builder
-    public CheckPoint(String content, Subtopic subtopic) {
-        this.content = content;
-        this.subtopic = subtopic;
-        this.createdAt = LocalDateTime.now();
-    }
-}
-
