--- src/main/java/org/dongguk/onroad/loadmap/domain/Quiz.java
@@ -1,50 +1,0 @@
-package org.dongguk.onroad.loadmap.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.dongguk.onroad.loadmap.domain.type.EQuizType;
-
-@Entity
-@Table(name = "quizzes")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Quiz {
-
-    /* -------------------------------------------- */
-    /* Default Column ----------------------------- */
-    /* -------------------------------------------- */
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    /* -------------------------------------------- */
-    /* Information Column ------------------------- */
-    /* -------------------------------------------- */
-    @Enumerated(EnumType.STRING)
-    @Column(name = "type", nullable = false)
-    private EQuizType type;
-
-    @Column(name = "content" ,length = 500, nullable = false)
-    private String content;
-
-    /* -------------------------------------------- */
-    /* One To One Mapping ------------------------- */
-    /* -------------------------------------------- */
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "check_point_id", nullable = false)
-    private CheckPoint checkPoint;
-
-    /* -------------------------------------------- */
-    /* Methods ------------------------------------ */
-    /* -------------------------------------------- */
-    @Builder
-    public Quiz(EQuizType type, String content, CheckPoint checkPoint) {
-        this.type = type;
-        this.content = content;
-        this.checkPoint = checkPoint;
-    }
-}
-
