--- src/main/java/org/dongguk/onroad/loadmap/domain/Section.java
@@ -1,61 +1,0 @@
-package org.dongguk.onroad.loadmap.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "sections")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Section {
-
-    /* -------------------------------------------- */
-    /* Default Column ----------------------------- */
-    /* -------------------------------------------- */
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    /* -------------------------------------------- */
-    /* Information Column ------------------------- */
-    /* -------------------------------------------- */
-    @Column(name = "week_index", nullable = false)
-    private Integer weekIndex;
-
-    @Column(name = "title", length = 200, nullable = false)
-    private String title;
-
-    @Column(name = "description", length = 500, nullable = false)
-    private String description;
-
-    /* -------------------------------------------- */
-    /* Timestamp Column --------------------------- */
-    /* -------------------------------------------- */
-    @Column(name = "created_at", nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    /* -------------------------------------------- */
-    /* Many To One Mapping ------------------------ */
-    /* -------------------------------------------- */
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "week_id", nullable = false)
-    private Week week;
-
-    /* -------------------------------------------- */
-    /* Methods ------------------------------------ */
-    /* -------------------------------------------- */
-    @Builder
-    public Section(Integer weekIndex, String title, String description, Week week) {
-        this.weekIndex = weekIndex;
-        this.title = title;
-        this.description = description;
-        this.week = week;
-        this.createdAt = LocalDateTime.now();
-    }
-}
-
