--- src/main/java/org/dongguk/onroad/loadmap/domain/Subtopic.java
@@ -1,57 +1,0 @@
-package org.dongguk.onroad.loadmap.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "subtopics")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Subtopic {
-
-    /* -------------------------------------------- */
-    /* Default Column ----------------------------- */
-    /* -------------------------------------------- */
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    /* -------------------------------------------- */
-    /* Information Column ------------------------- */
-    /* -------------------------------------------- */
-    @Column(name = "title", length = 100, nullable = false)
-    private String title;
-
-    @Column(name = "detail", length = 500, nullable = false)
-    private String detail;
-
-    /* -------------------------------------------- */
-    /* Timestamp Column --------------------------- */
-    /* -------------------------------------------- */
-    @Column(name = "created_at", nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    /* -------------------------------------------- */
-    /* Many To One Mapping ------------------------ */
-    /* -------------------------------------------- */
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "section_id", nullable = false)
-    private Section section;
-
-    /* -------------------------------------------- */
-    /* Methods ------------------------------------ */
-    /* -------------------------------------------- */
-    @Builder
-    public Subtopic(String title, String detail, Section section) {
-        this.title = title;
-        this.detail = detail;
-        this.section = section;
-        this.createdAt = LocalDateTime.now();
-    }
-}
-
