--- src/main/java/org/dongguk/onroad/loadmap/repository/SubtopicRepository.java
@@ -1,9 +1,0 @@
-package org.dongguk.onroad.loadmap.repository;
-
-import org.dongguk.onroad.loadmap.domain.Subtopic;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface SubtopicRepository extends JpaRepository<Subtopic, Long> {
-}
