+++ src/main/java/org/dongguk/onroad/roadmap/application/usecase/ReadLectureOverviewUseCase.java
@@ -1,0 +1,17 @@
+package org.dongguk.onroad.roadmap.application.usecase;
+
+import org.dongguk.onroad.core.annotation.bean.UseCase;
+import org.dongguk.onroad.roadmap.application.dto.request.ReadLectureOverviewResponseDto;
+
+import java.util.UUID;
+
+@UseCase
+public interface ReadLectureOverviewUseCase {
+
+    ReadLectureOverviewResponseDto execute(
+            Integer page,
+            Integer size,
+            UUID userId
+    );
+
+}
