+++ src/main/java/org/dongguk/onroad/roadmap/domain/Choice.java
@@ -1,0 +1,48 @@
+package org.dongguk.onroad.roadmap.domain;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "choices")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Choice {
+
+    /* -------------------------------------------- */
+    /* Default Column ----------------------------- */
+    /* -------------------------------------------- */
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    /* -------------------------------------------- */
+    /* Information Column ------------------------- */
+    /* -------------------------------------------- */
+    @Column(name = "content", length = 500, nullable = false)
+    private String content;
+
+    @Column(name = "is_answer", nullable = false)
+    private Boolean isAnswer;
+
+    /* -------------------------------------------- */
+    /* Many To One Mapping ------------------------ */
+    /* -------------------------------------------- */
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "quiz_id", nullable = false)
+    private Quiz quiz;
+
+    /* -------------------------------------------- */
+    /* Methods ------------------------------------ */
+    /* -------------------------------------------- */
+    @Builder
+    public Choice(String content, Boolean isAnswer, Quiz quiz) {
+        this.content = content;
+        this.isAnswer = isAnswer;
+        this.quiz = quiz;
+    }
+}
+
