+++ src/main/java/org/dongguk/onroad/roadmap/domain/Subtopic.java
@@ -1,20 +1,18 @@
-package org.dongguk.onroad.question.domain;
+package org.dongguk.onroad.roadmap.domain;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.dongguk.onroad.roadmap.domain.Lecture;
-import org.dongguk.onroad.security.domain.mysql.User;
 
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "questions")
+@Table(name = "subtopics")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Question {
+public class Subtopic {
 
     /* -------------------------------------------- */
     /* Default Column ----------------------------- */
@@ -26,18 +24,11 @@ public class Question {
     /* -------------------------------------------- */
     /* Information Column ------------------------- */
     /* -------------------------------------------- */
-    @Column(length = 500, nullable = false)
+    @Column(name = "title", length = 100, nullable = false)
     private String title;
 
-    @Lob
-    @Column(nullable = false)
-    private String content;
-
-    @Column(name = "img_url")
-    private String imgUrl;
-
-    @Column(name = "week_index", nullable = false)
-    private Integer weekIndex;
+    @Column(name = "detail", length = 500, nullable = false)
+    private String detail;
 
     /* -------------------------------------------- */
     /* Timestamp Column --------------------------- */
@@ -49,24 +40,17 @@ public class Question {
     /* Many To One Mapping ------------------------ */
     /* -------------------------------------------- */
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "lecture_id", nullable = false)
-    private Lecture lecture;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "student_id", nullable = false)
-    private User student;
+    @JoinColumn(name = "section_id", nullable = false)
+    private Section section;
 
     /* -------------------------------------------- */
     /* Methods ------------------------------------ */
     /* -------------------------------------------- */
     @Builder
-    public Question(String title, String content, String imgUrl, Integer weekIndex, Lecture lecture, User student) {
+    public Subtopic(String title, String detail, Section section) {
         this.title = title;
-        this.content = content;
-        this.imgUrl = imgUrl;
-        this.weekIndex = weekIndex;
-        this.lecture = lecture;
-        this.student = student;
+        this.detail = detail;
+        this.section = section;
         this.createdAt = LocalDateTime.now();
     }
 }
