+++ src/main/java/org/dongguk/onroad/roadmap/domain/UserChoice.java
@@ -1,0 +1,43 @@
+package org.dongguk.onroad.roadmap.domain;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.dongguk.onroad.security.domain.mysql.User;
+
+@Entity
+@Table(name = "user_choices")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class UserChoice {
+
+    /* -------------------------------------------- */
+    /* Default Column ----------------------------- */
+    /* -------------------------------------------- */
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    /* -------------------------------------------- */
+    /* Many To One Mapping ------------------------ */
+    /* -------------------------------------------- */
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "student_id", nullable = false)
+    private User student;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "choice_id", nullable = false)
+    private Choice choice;
+
+    /* -------------------------------------------- */
+    /* Methods ------------------------------------ */
+    /* -------------------------------------------- */
+    @Builder
+    public UserChoice(User student, Choice choice) {
+        this.student = student;
+        this.choice = choice;
+    }
+}
+
