+++ src/main/java/org/dongguk/onroad/roadmap/domain/UserLecture.java
@@ -1,0 +1,47 @@
+package org.dongguk.onroad.roadmap.domain;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.dongguk.onroad.security.domain.mysql.User;
+
+@Entity
+@Table(name = "user_lectures")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class UserLecture {
+
+    /* -------------------------------------------- */
+    /* Default Column ----------------------------- */
+    /* -------------------------------------------- */
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    /* -------------------------------------------- */
+    /* Many To One Mapping ------------------------ */
+    /* -------------------------------------------- */
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "student_id")
+    private User student;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "professor_id", nullable = false)
+    private User professor;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "lecture_id", nullable = false)
+    private Lecture lecture;
+
+    /* -------------------------------------------- */
+    /* Methods ------------------------------------ */
+    /* -------------------------------------------- */
+    @Builder
+    public UserLecture(User student, User professor, Lecture lecture) {
+        this.student = student;
+        this.professor = professor;
+        this.lecture = lecture;
+    }
+}
